rustc_attr
Function find_repr_attrs

pub fn find_repr_attrs(sess: &Session, attr: &Attribute) -> Vec<ReprAttr>

Parse #[repr(…)] forms.

Valid repr contents: any of the primitive integral type names (see int_type_of_word, below) to specify enum discriminant type; C, to use the same discriminant size that the corresponding C enum would or C structure layout, packed to remove padding, and transparent to delegate representation concerns to the only non-ZST field.