Crate rustc_attr

Functions and types dealing with attributes and meta items.

FIXME(Centril): For now being, much of the logic is still in rustc_ast::attr. The goal is to move the definition of MetaItem and things that don’t need to be in syntax to this crate.

Re-exports

pub use IntType::*;

pub use ReprAttr::*;

pub use StabilityLevel::*;

Modules

builtin 🔒

Parsing and validation of builtin attributes

session_diagnostics 🔒

Structs

AttrIdGenerator

Condition

ConstStability

Represents the #[rustc_const_unstable] and #[rustc_const_stable] attributes.

DefaultBodyStability

Represents the #[rustc_default_body_unstable] attribute.

Deprecation

MarkedAttrs

Stability

Represents the following attributes:

Enums

DeprecatedSince

Release in which an API is deprecated.

InlineAttr

InstructionSetAttr

IntType

OptimizeAttr

ReprAttr

StabilityLevel

The available stability levels.

StableSince

Rust release in which a feature is stabilized.

TransparencyError

UnstableReason

Constants

VERSION_PLACEHOLDER

The version placeholder that recently stabilized features contain inside the since field of the #[stable] attribute.

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Functions

allow_internal_unstable

cfg_matches

Tests if a cfg-pattern matches the cfg set

contains_name

eval_condition

Evaluate a cfg-like condition (with any and all), using eval to evaluate individual items.

filter_by_name

find_body_stability

Collects stability info from rustc_default_body_unstable attributes in attrs. Returns None if no stability attributes are found.

find_by_name

find_const_stability

Collects stability info from rustc_const_stable/rustc_const_unstable/rustc_promotable attributes in attrs. Returns None if no stability attributes are found.

find_crate_name

find_deprecation

Finds the deprecation attribute. None if none exists.

find_repr_attrs

Parse #[repr(…)] forms.

find_stability

Collects stability info from stable/unstable/rustc_allowed_through_unstable_modules attributes in attrs. Returns None if no stability attributes are found.

find_transparency

first_attr_value_str_by_name

is_builtin_attr

list_contains_name

mk_attr

mk_attr_from_item

mk_attr_name_value_str

mk_attr_nested_word

mk_attr_word

mk_doc_comment

parse_alignment

parse_confusables

Read the content of a rustc_confusables attribute, and return the list of candidate names.

parse_repr_attr

rustc_allow_const_fn_unstable